test: strip redundant integer casts from TEST_ASSERT_EQ arguments
TEST_ASSERT_EQ widens both arguments to int64_t internally, so outer
casts like (int), (int32_t), (uint32_t), etc. are redundant. A (int)
cast applied to a uint32_t enum silently truncates to 32-bit signed
before the widening, which can produce a false-passing comparison for
values >= 0x80000000.

Remove all such casts from 217 test files and 3 fuzz files using
scripts/utils/strip_assert_casts.py. The script only strips the leading
cast from each TEST_ASSERT_EQ argument; casts inside nested function
call arguments are left intact. A convergence loop handles cases where
multiple passes are needed due to nested call expressions.

Add scripts/utils/check_assert_casts.py as a pre-commit gate so future
test additions cannot re-introduce the pattern. Wire it into
scripts/git/pre-commit after the existing check_no_null block.

Also add missing @par MC/DC: declaration blocks to 63 test functions
across test_ra_flash.c, test_ra_fs_fat_variants.c, test_ra_mipi_phy.c,
test_ra_net_ipv4.c, test_ra_ptp.c, and test_ra_rsip_edge_cases.c, and
fix 549 bare NULL tokens (NULL -> nullptr per C23 policy) in 63 test
and fuzz files.

+++ tests/test_app_dac_waveform.c
@@ -57,12 +57,12 @@ static void test_dac_app_bringup_and_loop(void)
     .enable_channel0         = true,
     .enable_channel1         = false,
   };
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dac_b_init_configured(&cfg));
+  TEST_ASSERT_EQ(k_ra_ok, ra_dac_b_init_configured(&cfg));
   const uint16_t step =
     (uint16_t)((uint32_t)k_test_dac_app_max_code / (uint32_t)k_test_dac_app_steps);
   for (uint16_t i = 0U; i < (uint16_t)k_test_dac_app_steps; ++i) {
     const uint16_t code = (uint16_t)(i * step);
-    TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dac_b_write((uint8_t)k_test_dac_app_channel, code));
+    TEST_ASSERT_EQ(k_ra_ok, ra_dac_b_write((uint8_t)k_test_dac_app_channel, code));
   }
   TEST_END("dac_waveform: init + triangle write loop");
 }
@@ -115,10 +115,10 @@ static void test_dac_app_clamp(void)
     .enable_channel0         = true,
     .enable_channel1         = false,
   };
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dac_b_init_configured(&cfg));
+  TEST_ASSERT_EQ(k_ra_ok, ra_dac_b_init_configured(&cfg));
   TEST_ASSERT_EQ(
-    (int)k_ra_ok,
-    (int)ra_dac_b_write((uint8_t)k_test_dac_app_channel, (uint16_t)k_test_dac_app_oversize));
+    k_ra_ok,
+    ra_dac_b_write((uint8_t)k_test_dac_app_channel, (uint16_t)k_test_dac_app_oversize));
   TEST_END("dac_waveform: oversize value clamps OK");
 }
